delete side bar
